+++ kopi_docka/commands/advanced/__init__.py
@@ -40,6 +40,7 @@ from . import service_commands
 from . import system_commands
 from . import snapshot_commands
 from . import notification_commands
+from . import policy_commands
 
 # Register subcommand groups
 config_commands.register(admin_app)
@@ -48,3 +49,4 @@ service_commands.register(admin_app)
 system_commands.register(admin_app)
 snapshot_commands.register(admin_app)
 notification_commands.register(admin_app)
+policy_commands.register(admin_app)